PR-16 Refactor expense
