+++ app/src/main/java/com/pbs/expenseApp/data/AppContainer.kt
@@ -1,11 +1,11 @@
-package com.pbs.expenseApp.database
+package com.pbs.expenseApp.data
 
 import android.content.Context
 import com.pbs.expenseApp.data.repository.CategoryRepositoryImpl
+import com.pbs.expenseApp.data.repository.ExpenseRepositoryImpl
 import com.pbs.expenseApp.data.repository.UserRepositoryImpl
-import com.pbs.expenseApp.database.repositories.ExpenseRepository
-import com.pbs.expenseApp.database.repositories.offline.OfflineExpenseRepository
 import com.pbs.expenseApp.domain.repository.CategoryRepository
+import com.pbs.expenseApp.domain.repository.ExpenseRepository
 import com.pbs.expenseApp.domain.repository.UserRepository
 
 interface AppContainer {
@@ -22,6 +22,6 @@ class AppDataContainer(private val context: Context) : AppContainer {
         CategoryRepositoryImpl(ExpenseDatabase.getDatabase(context).categoryDao())
     }
     override val expenseRepository: ExpenseRepository by lazy {
-        OfflineExpenseRepository(ExpenseDatabase.getDatabase(context).expenseDao())
+        ExpenseRepositoryImpl(ExpenseDatabase.getDatabase(context).expenseDao())
     }
 }
