+++ app/src/main/java/com/pbs/expenseApp/data/ExpenseDatabase.kt
@@ -1,15 +1,15 @@
-package com.pbs.expenseApp.database
+package com.pbs.expenseApp.data
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import com.pbs.expenseApp.data.daos.CategoryDao
+import com.pbs.expenseApp.data.daos.ExpenseDao
 import com.pbs.expenseApp.data.daos.UserDao
-import com.pbs.expenseApp.database.converters.DateConverters
-import com.pbs.expenseApp.database.daos.ExpenseDao
-import com.pbs.expenseApp.database.entities.Expense
+import com.pbs.expenseApp.data.converters.DateConverters
 import com.pbs.expenseApp.domain.model.Category
+import com.pbs.expenseApp.domain.model.Expense
 import com.pbs.expenseApp.domain.model.User
 
 @Database(
