+++ app/src/main/java/com/pbs/expenseApp/data/daos/ExpenseDao.kt
@@ -1,13 +1,13 @@
-package com.pbs.expenseApp.database.daos
+package com.pbs.expenseApp.data.daos
 
-import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.pbs.expenseApp.database.entities.Expense
+import com.pbs.expenseApp.domain.model.Expense
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface ExpenseDao {
@@ -18,7 +18,7 @@ interface ExpenseDao {
     @Delete
     suspend fun delete(expense: Expense)
     @Query("SELECT * from expenses WHERE uuid = :uuid")
-    fun getExpense(uuid: String): LiveData<Expense?>
+    fun getExpense(uuid: String): Flow<Expense?>
     @Query("SELECT * from expenses")
-    fun getAllExpenses(): LiveData<List<Expense>>
+    fun getAllExpenses(): Flow<List<Expense>>
 }
