--- app/src/main/java/com/pbs/expenseApp/database/repositories/Expense.kt
@@ -1,12 +1,0 @@
-package com.pbs.expenseApp.database.repositories
-
-import androidx.lifecycle.LiveData
-import com.pbs.expenseApp.database.entities.Expense
-
-interface ExpenseRepository {
-    suspend fun insertExpense(expense: Expense)
-    suspend fun updateExpense(expense: Expense)
-    suspend fun deleteExpense(expense: Expense)
-    fun getExpenseStream(uuid: String): LiveData<Expense?>
-    fun getAllExpensesStream(): LiveData<List<Expense>>
-}
